polishing, doc and comments update
