+++ interpreter/shared/src/main/scala/sigmastate/eval/Extensions.scala
@@ -3,7 +3,7 @@ package sigmastate.eval
 import debox.cfor
 import org.ergoplatform.ErgoBox
 import org.ergoplatform.ErgoBox.TokenId
-import scorex.crypto.authds.avltree.batch.{Insert, Lookup, Remove, Update}
+import scorex.crypto.authds.avltree.batch.{Insert, InsertOrUpdate, Lookup, Remove, Update}
 import scorex.crypto.authds.{ADKey, ADValue}
 import scorex.util.encode.Base16
 import sigma.ast.SType.AnyOps
@@ -91,7 +91,7 @@ object Extensions {
         val bv = CAvlTreeVerifier(tree, proof)
         entries.forall { case (key, value) =>
           val insertRes = bv.performOneOperation(Insert(ADKey @@ key.toArray, ADValue @@ value.toArray))
-          if (insertRes.isFailure) {
+          if (insertRes.isFailure && !VersionContext.current.isV6SoftForkActivated) {
             syntax.error(s"Incorrect insert for $tree (key: $key, value: $value, digest: ${tree.digest}): ${insertRes.failed.get}}")
           }
           insertRes.isSuccess
@@ -112,6 +112,24 @@ object Extensions {
         val bv = CAvlTreeVerifier(tree, proof)
         operations.forall { case (key, value) =>
           bv.performOneOperation(Update(ADKey @@ key.toArray, ADValue @@ value.toArray)).isSuccess
+        }
+        bv.digest match {
+          case Some(d) => Some(tree.updateDigest(Colls.fromArray(d)))
+          case _ => None
+        }
+      }
+    }
+
+    def insertOrUpdate(
+                entries: Coll[(Coll[Byte], Coll[Byte])],
+                proof: Coll[Byte]): Option[AvlTree] = {
+      if (!tree.isInsertAllowed || !tree.isUpdateAllowed) {
+        None
+      } else {
+        val bv = CAvlTreeVerifier(tree, proof)
+        entries.forall { case (key, value) =>
+          val insertRes = bv.performOneOperation(InsertOrUpdate(ADKey @@ key.toArray, ADValue @@ value.toArray))
+          insertRes.isSuccess
         }
         bv.digest match {
           case Some(d) => Some(tree.updateDigest(Colls.fromArray(d)))
