+++ sc/shared/src/test/scala/sigma/LanguageSpecificationV5.scala
@@ -3174,11 +3174,6 @@ class LanguageSpecificationV5 extends LanguageSpecificationBase { suite =>
 
   type BatchProver = BatchAVLProver[Digest32, Blake2b256.type]
 
-  def performInsert(avlProver: BatchProver, key: Coll[Byte], value: Coll[Byte]) = {
-    avlProver.performOneOperation(Insert(ADKey @@ key.toArray, ADValue @@ value.toArray))
-    val proof = avlProver.generateProof().toColl
-    proof
-  }
 
   def performUpdate(avlProver: BatchProver, key: Coll[Byte], value: Coll[Byte]) = {
     avlProver.performOneOperation(Update(ADKey @@ key.toArray, ADValue @@ value.toArray))
@@ -3201,144 +3196,6 @@ class LanguageSpecificationV5 extends LanguageSpecificationBase { suite =>
   }
 
   type KV = (Coll[Byte], Coll[Byte])
-
-  property("AvlTree.insert equivalence") {
-    val insert = existingFeature((t: (AvlTree, (Coll[KV], Coll[Byte]))) => t._1.insert(t._2._1, t._2._2),
-      "{ (t: (AvlTree, (Coll[(Coll[Byte], Coll[Byte])], Coll[Byte]))) => t._1.insert(t._2._1, t._2._2) }",
-      FuncValue(
-        Vector(
-          (
-              1,
-              STuple(
-                Vector(
-                  SAvlTree,
-                  STuple(Vector(SCollectionType(STuple(Vector(SByteArray, SByteArray))), SByteArray))
-                )
-              )
-              )
-        ),
-        BlockValue(
-          Vector(
-            ValDef(
-              3,
-              List(),
-              SelectField.typed[Value[STuple]](
-                ValUse(
-                  1,
-                  STuple(
-                    Vector(
-                      SAvlTree,
-                      STuple(Vector(SCollectionType(STuple(Vector(SByteArray, SByteArray))), SByteArray))
-                    )
-                  )
-                ),
-                2.toByte
-              )
-            )
-          ),
-          MethodCall.typed[Value[SOption[SAvlTree.type]]](
-            SelectField.typed[Value[SAvlTree.type]](
-              ValUse(
-                1,
-                STuple(
-                  Vector(
-                    SAvlTree,
-                    STuple(Vector(SCollectionType(STuple(Vector(SByteArray, SByteArray))), SByteArray))
-                  )
-                )
-              ),
-              1.toByte
-            ),
-            SAvlTreeMethods.getMethodByName("insert"),
-            Vector(
-              SelectField.typed[Value[SCollection[STuple]]](
-                ValUse(
-                  3,
-                  STuple(Vector(SCollectionType(STuple(Vector(SByteArray, SByteArray))), SByteArray))
-                ),
-                1.toByte
-              ),
-              SelectField.typed[Value[SCollection[SByte.type]]](
-                ValUse(
-                  3,
-                  STuple(Vector(SCollectionType(STuple(Vector(SByteArray, SByteArray))), SByteArray))
-                ),
-                2.toByte
-              )
-            ),
-            Map()
-          )
-        )
-      ))
-
-    val testTraceBase = Array(
-      FixedCostItem(Apply),
-      FixedCostItem(FuncValue),
-      FixedCostItem(GetVar),
-      FixedCostItem(OptionGet),
-      FixedCostItem(FuncValue.AddToEnvironmentDesc, FixedCost(JitCost(5))),
-      ast.SeqCostItem(CompanionDesc(BlockValue), PerItemCost(JitCost(1), JitCost(1), 10), 1),
-      FixedCostItem(ValUse),
-      FixedCostItem(SelectField),
-      FixedCostItem(FuncValue.AddToEnvironmentDesc, FixedCost(JitCost(5))),
-      FixedCostItem(ValUse),
-      FixedCostItem(SelectField),
-      FixedCostItem(MethodCall),
-      FixedCostItem(ValUse),
-      FixedCostItem(SelectField),
-      FixedCostItem(ValUse),
-      FixedCostItem(SelectField),
-      FixedCostItem(SAvlTreeMethods.isInsertAllowedMethod, FixedCost(JitCost(15)))
-    )
-    val costDetails1 = TracedCost(testTraceBase)
-    val costDetails2 = TracedCost(
-      testTraceBase ++ Array(
-        ast.SeqCostItem(NamedDesc("CreateAvlVerifier"), PerItemCost(JitCost(110), JitCost(20), 64), 70),
-        ast.SeqCostItem(NamedDesc("InsertIntoAvlTree"), PerItemCost(JitCost(40), JitCost(10), 1), 1),
-        FixedCostItem(SAvlTreeMethods.updateDigestMethod, FixedCost(JitCost(40)))
-      )
-    )
-
-    forAll(keyCollGen, bytesCollGen) { (key, value) =>
-      val (tree, avlProver) = createAvlTreeAndProver()
-      val preInsertDigest = avlProver.digest.toColl
-      val insertProof = performInsert(avlProver, key, value)
-      val kvs = Colls.fromItems((key -> value))
-
-      { // positive
-        val preInsertTree = createTree(preInsertDigest, insertAllowed = true)
-        val input = (preInsertTree, (kvs, insertProof))
-        val (res, _) = insert.checkEquality(input).getOrThrow
-        res.isDefined shouldBe true
-        insert.checkExpected(input, Expected(Success(res), 1796, costDetails2, 1796, Seq.fill(4)(2102)))
-      }
-
-      { // negative: readonly tree
-        val readonlyTree = createTree(preInsertDigest)
-        val input = (readonlyTree, (kvs, insertProof))
-        val (res, _) = insert.checkEquality(input).getOrThrow
-        res.isDefined shouldBe false
-        insert.checkExpected(input, Expected(Success(res), 1772, costDetails1, 1772, Seq.fill(4)(2078)))
-      }
-
-      { // negative: invalid key
-        val tree = createTree(preInsertDigest, insertAllowed = true)
-        val invalidKey = key.map(x => (-x).toByte) // any other different from key
-        val invalidKvs = Colls.fromItems((invalidKey -> value)) // NOTE, insertProof is based on `key`
-        val input = (tree, (invalidKvs, insertProof))
-        val (res, _) = insert.checkEquality(input).getOrThrow
-        res.isDefined shouldBe true // TODO v6.0: should it really be true? (looks like a bug) (see https://github.com/ScorexFoundation/sigmastate-interpreter/issues/908)
-        insert.checkExpected(input, Expected(Success(res), 1796, costDetails2, 1796, Seq.fill(4)(2102)))
-      }
-
-      { // negative: invalid proof
-        val tree = createTree(preInsertDigest, insertAllowed = true)
-        val invalidProof = insertProof.map(x => (-x).toByte) // any other different from proof
-        val res = insert.checkEquality((tree, (kvs, invalidProof)))
-        res.isFailure shouldBe true
-      }
-    }
-  }
 
   property("AvlTree.update equivalence") {
     val update = existingFeature((t: (AvlTree, (Coll[KV], Coll[Byte]))) => t._1.update(t._2._1, t._2._2),
